prog/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
azure-core==1.23.1
azure-iot-device==2.11.0
azure-storage-blob==12.11.0
bcrypt==3.2.0
bokeh==2.3.3
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
cftime==1.5.2
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorcet==2.0.6
cryptography==3.3.1
cycler==0.10.0
dash==1.18.1
dash-auth==1.4.1
dash-bootstrap-components==0.10.7
dash-core-components==1.14.1
dash-extensions==0.0.38
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dask==2021.8.0
datashader==0.13.0
datashape==0.5.2
dateparser==1.0.0
deprecation==2.1.0
distributed==2021.8.0
et-xmlfile==1.1.0
ffmpy==0.3.0
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.8.0
Flask-SeaSurf==0.2.2
fsspec==2021.7.0
future==0.18.2
HeapDict==1.0.1
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
install==1.3.5
isodate==0.6.1
itsdangerous==1.1.0
janus==1.0.0
Jinja2==2.11.2
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
MarkupSafe==1.1.1
matplotlib==3.3.3
more-itertools==8.6.0
msgpack==1.0.2
msrest==0.6.21
multipledispatch==0.6.0
netCDF4==1.5.8
numba==0.53.1
numpy==1.22.2
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.0
paho-mqtt==1.6.1
pandas==1.3.5
param==1.11.1
paramiko==2.7.2
partd==1.2.0
Pillow==8.0.1
pip==22.0.4
plotly==4.14.1
pluggy==0.13.1
psutil==5.9.0
py==1.10.0
pycparser==2.20
pyct==0.4.8
PyNaCl==1.4.0
pyodbc==4.0.30
pyparsing==2.4.7
pyproj==3.3.0
pyserial==3.5
pysftp==0.2.9
pyshorteners==1.0.1
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
retrying==1.3.3
scipy==1.5.4
six==1.15.0
socks==0
sortedcontainers==2.4.0
tblib==1.7.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
typing_extensions==4.1.1
tzlocal==2.1
ua-parser==0.10.0
urllib3==1.26.9
Werkzeug==1.0.1
wget==3.2
xarray==0.19.0
zict==2.0.0
